getLost/requirements.txt
asgiref==3.2.7
autopep8==1.5.2
Babel==2.8.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.3
chardet==4.0.0
cryptography==3.1.1
defusedxml==0.6.0
Django==3.0.6
django-admin-interface==0.15.3
django-allauth==0.43.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.2.0
django-colorfield==0.4.1
django-countries==7.0
django-crispy-forms==1.9.1
django-currencies==0.10.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-modeladmin-reorder==0.3.1
django-money==1.3.1
django-phonenumber-field==4.0.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
fanstatic==1.1
idna==2.10
js.momentjs==2.13.1
jsonfield==2.1.1
nodejs==0.1.1
oauthlib==3.1.0
optional-django==0.1.0
phonenumbers==8.12.4
Pillow==7.1.2
psycopg2==2.8.5
py-moneyed==0.8.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.25.1
requests-oauthlib==1.3.0
shutilwhich==1.1.0
six==1.15.0
soupsieve==2.2
sqlparse==0.3.1
stripe==2.56.0
toml==0.10.1
urllib3==1.26.3
WebOb==1.8.6
